# config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Bangladesh E-commerce API",
    "version": "1.0.0",
    "description": "Production-ready e-commerce API with authentication, products, cart, and orders"
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "User registration",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "email", "password", "phone"],
              "properties": {
                "name": { "type": "string" },
                "email": { "type": "string", "format": "email" },
                "password": { "type": "string", "format": "password" },
                "phone": { "type": "string" },
                "address": { "type": "string" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Registration successful" },
          "400": { "description": "Validation error or email exists" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "password"],
              "properties": {
                "email": { "type": "string", "format": "email" },
                "password": { "type": "string", "format": "password" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Login successful" },
          "400": { "description": "Invalid credentials" }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get all products",
        "parameters": [
          { "name": "page", "in": "query", "type": "integer" },
          { "name": "limit", "in": "query", "type": "integer" },
          { "name": "category", "in": "query", "type": "string" }
        ],
        "responses": {
          "200": { "description": "List of products" }
        }
      }
    },
    "/cart": {
      "get": {
        "summary": "Get user cart",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Cart items" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Add to cart",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["product_id"],
              "properties": {
                "product_id": { "type": "integer" },
                "quantity": { "type": "integer" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "Product added" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "http",
      "scheme": "bearer",
      "bearerFormat": "JWT"
    }
  }
}